fix(frontend): preserve upload error fallbacks and split turtle API module
